+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### Changes
+
+* Default `chromedriver` version is latest, not fixed by constant
+
 ### 0.13.0 (2021-04-15)
 
 ### New Features